+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [3.1.0]
 ## Added
 - Exit button to installing content dialog
 - Race button to event view